Move UpsellBadge into the toolbar folder
After all, it's part of the toolbar.
