+++ src/app/components/client/toolbar/UpsellBadge.module.scss
@@ -1,4 +1,4 @@
-@import "../../tokens";
+@import "../../../tokens";
 
 .upsellBadge {
   align-items: center;
