+++ src/app/components/client/toolbar/UpsellDialog.module.scss
@@ -1,4 +1,4 @@
-@import "../../tokens";
+@import "../../../tokens";
 
 .modalContent {
   display: flex;
@@ -17,7 +17,7 @@
     }
 
     dd {
-      background: url("../../(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/fix/images/icon-check-mark.svg")
+      background: url("../../../(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/fix/images/icon-check-mark.svg")
         no-repeat left top;
       list-style: none;
       margin: 0;
